+++ server/src/main/java/org/opensearch/index/query/MatchQueryBuilder.java
@@ -209,8 +209,10 @@ public class MatchQueryBuilder extends AbstractQueryBuilder<MatchQueryBuilder> {
     }
 
     /** Sets the fuzziness used when evaluated to a fuzzy query type. Defaults to "AUTO". */
-    public MatchQueryBuilder fuzziness(Object fuzziness) {
-        this.fuzziness = Fuzziness.build(fuzziness);
+    public MatchQueryBuilder fuzziness(Fuzziness fuzziness) {
+        if (fuzziness != null) {
+            this.fuzziness = fuzziness;
+        }
         return this;
     }
 
@@ -565,9 +567,7 @@ public class MatchQueryBuilder extends AbstractQueryBuilder<MatchQueryBuilder> {
         matchQuery.operator(operator);
         matchQuery.analyzer(analyzer);
         matchQuery.minimumShouldMatch(minimumShouldMatch);
-        if (fuzziness != null) {
-            matchQuery.fuzziness(fuzziness);
-        }
+        matchQuery.fuzziness(fuzziness);
         matchQuery.fuzzyRewrite(fuzzyRewrite);
         matchQuery.prefixLength(prefixLength);
         matchQuery.fuzzyTranspositions(fuzzyTranspositions);